refactor(applications-service-api): combine document services into one 'ni' service
