+++ packages/applications-service-api/__tests__/unit/services/document.ni.service.test.js
@@ -1,7 +1,7 @@
 const {
 	fetchDocuments,
 	getAvailableFilters
-} = require('../../../src/services/document.v3.service');
+} = require('../../../src/services/document.ni.service');
 const { DB_DOCUMENTS, DB_FILTERS } = require('../../__data__/documents');
 const { Op } = require('sequelize');
